17_Multivariate_Time_Series_Forecasting_using_Python/requirements.txt
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
statsmodels>=0.14.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
